Refactor tests into modular structure
